botsautos met gegevens

+++ src/nl/youngcapital/kermis/attracties/Attractie.java
@@ -3,7 +3,7 @@ package nl.youngcapital.kermis.attracties;
 public class Attractie {
 
    public String attractieNaam;
-   public int prijs;
+   public double prijs;
    public int oppervlakte;
 
    public Attractie() {
